:bug: autolock fix
Remove hiding screen when coming back from inactive state.
